+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The focus of this release has been on making HashMap much more complete.
 * Added `HashMap.has()`
 * Added `HashMap.get()`
 * Added `HashMap.size()`
+* Added `HashMap.values()`
 * Added `deleteProperty()`
 
 ### Fixes